backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
stripe
psycopg2-binary
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
openai
requests
jinja2
weasyprint
pytest
